Move QueryHandler and RecordManager handlers into zeroconf.handlers
